+++ voicevox_engine/util/__init__.py
@@ -1,0 +1,6 @@
+from decode_base64_waves import DecodeBase64WavesException, decode_base64_waves
+
+__all__ = [
+    "decode_base64_waves",
+    "DecodeBase64WavesException",
+]
